feat: provide extraParams and tls config for db
This change provides the user with the ability to configure addition dsn
parameters for the db connection that pipeline server makes to the
provided sql server.

The external db connection will also default to making a tls enabled
connection, users looking to disable this will need to provide custom
parameters.

Signed-off-by: Humair Khan <HumairAK@users.noreply.github.com>